mv create response message func -> create_response_message.cpp
