+++ srcs/HttpResponse/GET/get_request_body.cpp
@@ -36,26 +36,6 @@ bool HttpResponse::is_cgi_file() const {
 }
 
 
-void HttpResponse::get_error_page_to_body() {
-    this->body_buf_.clear();
-
-    // get_error_page_path
-    DEBUG_PRINT(CYAN, "  get_error_page 1 target: %s, request_status: %d",
-                this->request_.target().c_str(), this->status_code());
-    Result<std::string, int> result;
-    result = Config::get_error_page_path(this->server_config_,
-                                         this->request_.target(),
-                                         this->status_code());
-    if (result.is_err()) {
-        DEBUG_PRINT(CYAN, "  get_error_page 2 -> err");
-        return;
-    }
-    DEBUG_PRINT(CYAN, "  get_error_page 3");
-    std::string error_page_path = result.ok_value();
-    DEBUG_PRINT(CYAN, "  get_error_page 4 error_page_path: %s", error_page_path.c_str());
-
-    get_file_content(error_page_path, &this->body_buf_);
-}
 
 
 bool HttpResponse::is_redirect() const {
